# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medcom API",
    "description": "Medcom API documentation",
    "termsOfService": "",
    "contact": {
      "name": "",
      "url": "",
      "email": ""
    },
    "license": {
      "name": "unlicensed",
      "url": ""
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/admin/users/{login}": {
      "get": {
        "tags": ["user-resource"],
        "operationId": "getUser",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["user-resource"],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["user-resource"],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/register-pharmacy": {
      "post": {
        "tags": ["account-resource"],
        "operationId": "registerPharmacy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmacyRegistrationVM"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/account": {
      "get": {
        "tags": ["account-resource"],
        "operationId": "getAccount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["account-resource"],
        "operationId": "saveAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/account/reset-password/init": {
      "post": {
        "tags": ["account-resource"],
        "operationId": "requestPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/account/reset-password/finish": {
      "post": {
        "tags": ["account-resource"],
        "operationId": "finishPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyAndPasswordVM"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/account/change-password": {
      "post": {
        "tags": ["account-resource"],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/activate": {
      "get": {
        "tags": ["account-resource"],
        "operationId": "activateAccount",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminUserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
          },
          "firstName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5
          },
          "imageUrl": {
            "type": "string",
            "maxLength": 256,
            "minLength": 0
          },
          "activated": {
            "type": "boolean"
          },
          "langKey": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2
          },
          "createdBy": {
            "type": "string",
            "readOnly": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "lastModifiedBy": {
            "type": "string",
            "readOnly": true
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "pharmacyId": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "required": ["login"]
      },
      "PharmacyRegistrationVM": {
        "type": "object",
        "properties": {
          "pharmacyName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "address": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "pharmacyEmail": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "website": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "adminLogin": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$"
          },
          "adminPassword": {
            "type": "string",
            "maxLength": 100,
            "minLength": 4
          },
          "adminFirstName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "adminLastName": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "adminEmail": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5
          },
          "langKey": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2
          }
        },
        "required": ["adminEmail", "adminLogin", "adminPassword", "pharmacyName"]
      },
...
      "KeyAndPasswordVM": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
      "PasswordChangeDTO": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      }
    }
  }
}