# src/main/java/com/zidir/medcom/web/rest/AccountResource.java
package com.zidir.medcom.web.rest;

import com.zidir.medcom.domain.Pharmacy;
import com.zidir.medcom.domain.User;
import com.zidir.medcom.repository.UserRepository;
import com.zidir.medcom.security.SecurityUtils;
import com.zidir.medcom.service.MailService;
import com.zidir.medcom.service.UserService;
import com.zidir.medcom.service.dto.AdminUserDTO;
import com.zidir.medcom.service.dto.PasswordChangeDTO;
import com.zidir.medcom.web.rest.errors.*;
import com.zidir.medcom.web.rest.vm.KeyAndPasswordVM;
import com.zidir.medcom.web.rest.vm.ManagedUserVM;
import com.zidir.medcom.web.rest.vm.PharmacyRegistrationVM;
import jakarta.validation.Valid;
import java.util.*;
import org.apache.commons.lang3.StringUtils;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

/**
 * REST controller for managing the current user's account.
 */
@RestController
@RequestMapping("/api")
public class AccountResource {

    private static class AccountResourceException extends RuntimeException {

        private AccountResourceException(String message) {
            super(message);
        }
    }

    private static final Logger LOG = LoggerFactory.getLogger(AccountResource.class);

    private final UserRepository userRepository;

    private final UserService userService;

    private final MailService mailService;

    public AccountResource(UserRepository userRepository, UserService userService, MailService mailService) {
        this.userRepository = userRepository;
        this.userService = userService;
        this.mailService = mailService;
    }

    /**
     * {@code POST  /register} : register the user.
     * DISABLED: Public registration is disabled. Users must be created by pharmacy admins.
     *
     * @param managedUserVM the managed user View Model.
     * @throws InvalidPasswordException {@code 400 (Bad Request)} if the password is incorrect.
     * @throws EmailAlreadyUsedException {@code 400 (Bad Request)} if the email is already used.
     * @throws LoginAlreadyUsedException {@code 400 (Bad Request)} if the login is already used.
     */
    /*
    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    public void registerAccount(@Valid @RequestBody ManagedUserVM managedUserVM) {
        if (isPasswordLengthInvalid(managedUserVM.getPassword())) {
            throw new InvalidPasswordException();
        }
        User user = userService.registerUser(managedUserVM, managedUserVM.getPassword());
        mailService.sendActivationEmail(user);
    }
    */

    /**
     * {@code POST  /register-pharmacy} : register a pharmacy with its admin user.
     *
     * @param pharmacyRegistrationVM the pharmacy registration data.
     * @throws InvalidPasswordException {@code 400 (Bad Request)} if the password is incorrect.
     * @throws EmailAlreadyUsedException {@code 400 (Bad Request)} if the email is already used.
     * @throws LoginAlreadyUsedException {@code 400 (Bad Request)} if the login is already used.
     */
    @PostMapping("/register-pharmacy")
    @ResponseStatus(HttpStatus.CREATED)
    public void registerPharmacy(@Valid @RequestBody PharmacyRegistrationVM pharmacyRegistrationVM) {
        if (isPasswordLengthInvalid(pharmacyRegistrationVM.getAdminPassword())) {
            throw new InvalidPasswordException();
        }

        // Check if login already exists
        userRepository
            .findOneByLogin(pharmacyRegistrationVM.getAdminLogin().toLowerCase())
            .ifPresent(existingUser -> {
                throw new LoginAlreadyUsedException();
            });

        // Check if email already exists
        userRepository
            .findOneByEmailIgnoreCase(pharmacyRegistrationVM.getAdminEmail())
            .ifPresent(existingUser -> {
                throw new EmailAlreadyUsedException();
            });

        // Create pharmacy entity
        Pharmacy pharmacy = new Pharmacy();
        pharmacy.setName(pharmacyRegistrationVM.getPharmacyName());
        pharmacy.setAddress(pharmacyRegistrationVM.getAddress());
        pharmacy.setEmail(pharmacyRegistrationVM.getPharmacyEmail());
        pharmacy.setPhone(pharmacyRegistrationVM.getPhone());
        pharmacy.setWebsite(pharmacyRegistrationVM.getWebsite());
        pharmacy.setActive(true);

        // Create admin user DTO
        AdminUserDTO adminUserDTO = new AdminUserDTO();
        adminUserDTO.setLogin(pharmacyRegistrationVM.getAdminLogin());
        adminUserDTO.setFirstName(pharmacyRegistrationVM.getAdminFirstName());
        adminUserDTO.setLastName(pharmacyRegistrationVM.getAdminLastName());
        adminUserDTO.setEmail(pharmacyRegistrationVM.getAdminEmail());
        adminUserDTO.setLangKey(pharmacyRegistrationVM.getLangKey());
        adminUserDTO.setActivated(true);

        // Register pharmacy with admin
        User adminUser = userService.registerPharmacyWithAdmin(pharmacy, adminUserDTO, pharmacyRegistrationVM.getAdminPassword());
        LOG.debug("Pharmacy and admin user created successfully: pharmacy={}, admin={}", pharmacy.getName(), adminUser.getLogin());
    }

    /**
     * {@code GET  /activate} : activate the registered user.
     *
     * @param key the activation key.
     * @throws RuntimeException {@code 500 (Internal Server Error)} if the user couldn't be activated.
     */
    @GetMapping("/activate")
    public void activateAccount(@RequestParam(value = "key") String key) {
        Optional<User> user = userService.activateRegistration(key);
        if (!user.isPresent()) {
            throw new AccountResourceException("No user was found for this activation key");
        }
    }

    /**
     * {@code GET  /account} : get the current user.
     *
     * @return the current user.
     * @throws RuntimeException {@code 500 (Internal Server Error)} if the user couldn't be returned.
     */
    @GetMapping("/account")
    public AdminUserDTO getAccount() {
        return userService
            .getUserWithAuthorities()
            .map(AdminUserDTO::new)
            .orElseThrow(() -> new AccountResourceException("User could not be found"));
    }

    /**
     * {@code POST  /account} : update the current user information.
     *
     * @param userDTO the current user information.
     * @throws EmailAlreadyUsedException {@code 400 (Bad Request)} if the email is already used.
     * @throws RuntimeException {@code 500 (Internal Server Error)} if the user login wasn't found.
     */
    @PostMapping("/account")
    public void saveAccount(@Valid @RequestBody AdminUserDTO userDTO) {
        String userLogin = SecurityUtils.getCurrentUserLogin()
            .orElseThrow(() -> new AccountResourceException("Current user login not found"));
        Optional<User> existingUser = userRepository.findOneByEmailIgnoreCase(userDTO.getEmail());
        if (existingUser.isPresent() && (!existingUser.orElseThrow().getLogin().equalsIgnoreCase(userLogin))) {
            throw new EmailAlreadyUsedException();
        }
        Optional<User> user = userRepository.findOneByLogin(userLogin);
        if (!user.isPresent()) {
            throw new AccountResourceException("User could not be found");
        }
        userService.updateUser(
            userDTO.getFirstName(),
            userDTO.getLastName(),
            userDTO.getEmail(),
            userDTO.getLangKey(),
            userDTO.getImageUrl()
        );
    }

    /**
     * {@code POST  /account/change-password} : changes the current user's password.
     *
     * @param passwordChangeDto current and new password.
     * @throws InvalidPasswordException {@code 400 (Bad Request)} if the new password is incorrect.
     */
    @PostMapping(path = "/account/change-password")
    public void changePassword(@RequestBody PasswordChangeDTO passwordChangeDto) {
        if (isPasswordLengthInvalid(passwordChangeDto.getNewPassword())) {
            throw new InvalidPasswordException();
        }
        userService.changePassword(passwordChangeDto.getCurrentPassword(), passwordChangeDto.getNewPassword());
    }

    /**
     * {@code POST   /account/reset-password/init} : Send an email to reset the password of the user.
     *
     * @param mail the mail of the user.
     */
    @PostMapping(path = "/account/reset-password/init")
    public void requestPasswordReset(@RequestBody String mail) {
        Optional<User> user = userService.requestPasswordReset(mail);
        if (user.isPresent()) {
            mailService.sendPasswordResetMail(user.orElseThrow());
        } else {
            // Pretend the request has been successful to prevent checking which emails really exist
            // but log that an invalid attempt has been made
            LOG.warn("Password reset requested for non existing mail");
        }
    }

    /**
     * {@code POST   /account/reset-password/finish} : Finish to reset the password of the user.
     *
     * @param keyAndPassword the generated key and the new password.
     * @throws InvalidPasswordException {@code 400 (Bad Request)} if the password is incorrect.
     * @throws RuntimeException {@code 500 (Internal Server Error)} if the password could not be reset.
     */
    @PostMapping(path = "/account/reset-password/finish")
    public void finishPasswordReset(@RequestBody KeyAndPasswordVM keyAndPassword) {
        if (isPasswordLengthInvalid(keyAndPassword.getNewPassword())) {
            throw new InvalidPasswordException();
        }
        Optional<User> user = userService.completePasswordReset(keyAndPassword.getNewPassword(), keyAndPassword.getKey());

        if (!user.isPresent()) {
            throw new AccountResourceException("No user was found for this reset key");
        }
    }

    private static boolean isPasswordLengthInvalid(String password) {
        return (
            StringUtils.isEmpty(password) ||
            password.length() < ManagedUserVM.PASSWORD_MIN_LENGTH ||
            password.length() > ManagedUserVM.PASSWORD_MAX_LENGTH
        );
    }
}